Arguments to pass to a Shiny App in a child process

runApp_args: Arguments that populate a runApp call that will be run in a child process.

golem_args: Arguments that populate an app run via the golem package.

The command is appended predefined commands and sent to a process object.

runApp_args(
  appDir = getwd(),
  test_port = httpuv::randomPort(),
  test_ip = getOption("shiny.host", "127.0.0.1"),
  test_path = tempdir(),
  test_trace = FALSE
)

golem_args(
  package_name = "",
  test_port = httpuv::randomPort(),
  test_ip = getOption("shiny.host", "127.0.0.1"),
  test_path = tempdir(),
  test_trace = FALSE
)

Arguments

appDir
The application to run. Should be one of the following (Default: getwd()):

A directory containing server.R, plus, either ui.R or a www directory that contains the file index.html.

A directory containing app.R.

An .R file containing a Shiny application, ending with an expression that produces a Shiny app object.

test_port
integer, port to run the app on. Default: httpuv::randomPort()

test_ip
The IPv4 address that the application should listen on.

test_path
character, Path the child process will have access to on the master, Default: tempdir()

test_trace
logical, turn on the shiny.trace option in the background proccess?. Default: FALSE

package_name
name of the golem package

Value

character

See also

Examples

runApp_args()

#> [1] "pkgload::load_all()"                                                                                                    
#> [2] "library(whereami)"                                                                                                      
#> [3] "whereami::set_whereami_log('/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpGWDxKK/reactor')"                     
#> [4] "options('shiny.port'= 11336, shiny.host='127.0.0.1', shiny.trace = FALSE)"                                              
#> [5] "shiny::runApp(appDir = '/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGWDxKK/file642238b0e0b/reference')"

golem_args()

#> [1] "pkgload::load_all()"                                                                               
#> [2] "library(whereami)"                                                                                 
#> [3] "whereami::set_whereami_log('/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpGWDxKK/reactor')"
#> [4] "options('shiny.port'= 47787, shiny.host='127.0.0.1', shiny.trace = FALSE)"                         
#> [5] "run_app()"